[TransactionEffects/API] Object Changes
Expose object changes directly on `TransactionEffectsAPI`, and
implement `ObjectChange` in the GraphQL API in terms of that.

In order to do this, the `ObjectChange` returned by
`TransactionEffectsAPI` needs to be a lowest common denominator
between Effects V1 and V2. This means that it does not include the
digests for input objects, because Effects V1 does not have this
information.

This concession is fine for the purposes of the GraphQL API, because
we do not rely on the digest there, and if someone wants to do a
verified read, they are free to confirm the digest from the contents
of the object against the digest in the raw transaction effects (for
v2).

This change also introduces a `location` field to GraphQL's
`ObjectChange`.  This is mainly to handle the case of an "unwrapped
then deleted" object, which will have neither an input state nor an
output state, but whose id was deleted.  For this to make sense, we
will need some other way to supply the ID being deleted.

Test Plan:

Tested via GraphQL E2E tests:

```
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```