[GraphQL/Object] Make `StoredObject` optional
Make it possible to create an `Object` type without a `StoredObject`.
This allows us to create objects that have not been indexed.  This
will be used when representing a `GenesisTransaction` which stores
object contents inline.

This pattern is also applicable for representing
`TransactionBlockEffects` that have not been indexed yet.

Test Plan:

Existing tests (this PR does not use the code path where an object is
created without its indexed counterpart, it only introduces the
possibility of that, to be used in a later PR).

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```